Add a sanity-checking initial sqitch change
Check that all the tables we expect to exist after bootstrapping do
exist and have required permissions, and that a few tables we
specifically expect to not exist, based on some of the last
pre-publication phases of development in the Mullins Lab, do not exist.
This is here as a sanity check to improve the chances that no sqitch
changes are run against a database that is in an unexpected state.