Add text on when it's ok to not upload chromats
Now that we're drawing a little more attention to this feature with the
'set all' button I want to make sure there's in-app guidance on when to
use it. Also moves the option to be last within the panel for each
product because that better suits its importance. The downside of that
is it makes the control more distant from the file-list area but I don't
think it's a big deal.