Require ViroDB Scientists in can_revise_sequence
Previously this method was written to allow calling it on a
Viroverse::Model::scientist as well, but in the current codebase it
never is, because it's called only on the logged-in scientist, which is
only ever set to a ViroDB::Result::Scientist. Validates the parameter
instead of just dying when the role predicates don't exist because why
not use the tools we've built for ourselves.