Add a date censoring machine
Like other single-purpose view-models, this is designed to work as a TT
plugin or not, encapsulate a single stupid function that doesn't really
want to be decomposed or have its logic distributed, fill in other
reasons why I like this pattern here.

The test tests that we get some expected answers. It's kind of a
fraudulent test because I sort of guessed at how the math would work out
and changed the expectation when I was wrong, but I guess it still works
as a regression test. Specifying and implementing an entire method of
calculating weeks, months, or years of duration that makes more human
sense than what I've managed to do so far is out of scope.