Censor dates in sample search results
The method here is to unconditionally add a `relative_date` to the hash
representation (and thus JSON representation, and thus representation in
search results) of a DistinctSampleSearch instance. This is
theoretically a poor choice, in that the DateCensor will do a couple of
extra queries for each sample displayed in the search results, which are
otherwise drawn from the single distinct_sample_search query, but these
are indexed lookups that should always be fast and this design overall
is more disposable than adding logic for relative dates to one database
view or another.

Displays absolute dates for derived samples, even to users getting the
censored UX, which seems sensible since those are not
clinically-relevant dates (in our world, anyway).

NB that as of this commit the search results are ordered by actual date
and thus leak information that they probably shouldn't to consider them
truly 'censored', but our use-case for this feature will involve data
where dates have all been rescaled to start at the same origin and so
the relative and absolute orderings will be identical there. The work on
this branch isn't meant to guarantee that no absolute dates will reach
the client, merely avoid doing so.