Refactor input/pcr_add controller action
It's not shorter but I've dramatically improved organization and
intelligibility, which in addition to being good on its own should
unblock some front-end development because it should be easier to make
incremental changes in concert with the backend.

Note for instance that the chunk of code parsing form parameters into
data structures and the chunk adding rows to the database are now
separate and the whole operation has a much more explicit looping
structure.

Adds an override for PolymeraseChainReactionTemplate::new to allow using
input_product as a generic parameter to the constructor and infer the
actual foreign key relationship on the model side. The inlined
constructor from __PACKAGE__->meta->make_immutable didn't pick up the
overridden `new` so the constructor inlining is disabled. This seems to
confuse dbicdump, and maybe there's a better way to solve this mismatch,
but I do not know it offhand.

More of this code should potentially be moved out of the controller
someday but on the other hand at least here we can keep an eye on it.