Retire the separate schema for derivation tables
The derivation and (derivation) protocol tables were originally created
in a separate Pg schema named "delta", partly for project organization
purposes and partly to to allow reuse of the name "protocol". This has
outlived whatever usefulness it had, and it makes sense to rename the
delta.protocol table to just be named "derivation_protocol". So this
sqitch change does all that and renames other objects like sequences and
indexes to match.

The distinct_sample_search materialized view doesn't require
modification, because the normal form the view query is converted to by
the system updates source table names when they're changed and aliases
columns from the source tables by position, so as long as column types
and order aren't changing (which they aren't) the view keeps working.
Hooray!