Change configuration interface to use Config::Any
This replaces the `Viroverse::config` and `Viroverse::config_local` modules with a simpler configuration mechanism that allows the site to put one or more configuration files in the Viroverse root directory in any format supported by `Config::Any`, named `viroverse_local.*` or `viroverse.*` (with the former taking precedence over the latter). It still provides default configuration as before.

Since it does a hash-level merge, overriding part of a hash is now possible. For example, with this in a viroverse_local.json:

```json
{
    "storage": "/home/viroverse/vagrant-storage",
    "features" : {
        "censor_dates": true
    }
}
```

this will enable the censor_dates feature while keeping the other configured features as the defaults.