Rework sequence deletion
• Doing sequence deletion with a javascript widget that sends a DELETE
XHR adds nothing to the user experience here. I'll keep exploiting
Angular to disable the submit button when the reason field is empty, but
get rid of the other Angular stuff.

• Unrelatedly, having the deletion-marking routine defined in the model
class is an idiosyncratic factoring that doesn't make sense to me at
this point. We don't need to do this deletion "from" anywhere other than
from clicking this button to submit to this controller so just handle it
in the controller. If ever that changes, well, we can move the code
around again.