Revamp contents of homepage
The project panels aren't super useful these days, because
project-management functionality is underused and some new users
confused projects with cohorts, causing discoverability issues. Instead
list the cohorts on the homepage, and if a user has no samples assigned
just remove that chunk of content entirely. The project panels are
relocated to the project index.