Fix execution of quality command-line tool
After revamping the configuration file mechanism, the interpolation of
this shell command became incorrect, and quality would not execute.
Instead we can see this telltale error in the log:

  sh: Viroverse::Config-: command not found

It works now. I spent zero time looking for other ways to improve this
code just now; it should probably do things like checking the shell
command for errors and handling them more sensibly. But a lot of this
should be more sensible than it is.