Add an installation script for production use
This script is intended to do setup tasks for the application that don't
require root privileges on the system: primarily Perl dependencies and a
couple other environmental things. Most of the parts are extracted from
devel/vagrant/provision, which remains in place for the time being, but
provisioning for Vagrant should probably be rewritten to use this
instead, with knobs added as necessary.