Fix revert script for sequence note migration
This is almost completely academic, but bit me when rolling forward and
back while working on a rework to add some database configuration
configurability. The note column was not added back to na_sequence in
the revert script; of course the deploy script expects the note column
to be there, so a forward-back-forward of this change wouldn't work.