Segregate listing of retired scientists
In our instance we have more of these than active users, and it makes
administration via these sorta barebones pages unwieldy/harder to
browse. With this change, also adjust how scientists are sorted since we
no longer have to exploit the sort order of the role names to
deemphasize the retired users.