Add numeric assay results for specific samples
Comes with a protocol table that establishes the unit and dbic-dumping

This also removes the sample_id column from lab_result_num and
lab_result_cat, where it has remained despite a lengthy deprecation
period. There's no reason to try to make sense of whatever data is left
in those columns in production at this point. Away it goes.