Allow adding PacBio sequences to arbitrary PCR
The safest way of finding PCR products that can reasonably have PacBio
sequencing added is through the PacBio pool search, which finds preps
that use our lab's pooling workflow. But we also want to minimally
support adding PacBio sequences to other arbitrary PCRs, and to be
expedient we're exploiting the PCR product finder this rather than
implementing another new search (which would be harder to design).

Other ingredients of this:

* Add the MolecularProduct role to models in the ViroDB world
* Emulate input_sample from that role on PacbioPool model
* Use a super secure and secret param to make sure we only are allowed
  to add PacBio sequences to arbitrary PCRs when asking nicely.