Replace usage of a pointless query wrapper
Instead of porting this weird data structure pattern to a new TissueType
result set I'm just repeating the effect in the two callsites, so
they're each easier to delete someday. It'd probably be trivial to
further refactor whatever consumes the data from these pages to work
with a simple ->order_by("name")->all but I'm keeping this minimal.