Stop using Viroverse::patient in Model::aliquot
Was used for the all_names query, which I could recreate more carefully
but I'm just replacing it here with the patient's highest priority name.
The lack of aliases in the affected view
(/freezer/search_freezers/aliquot_summary_by_box) is not going to bother
anyone. Further refactoring avoided--there's no reason to
put the single name in an array, but I don't want to touch more code.