Work-in-progress on ISLA importer
This is a wip state of an initial effort on an ISLA importer for the
Frenkel lab, pushed in case my time runs out before I'm able to finish
the train of thought I started here.

The general idea is to create an Import class of the existing model,
based on some configured job-global parameters and processing the rows
of a rectangular spreadsheet/CSV file. Then, to support uploading ISLA
worksheet templates, probably wrapping this importer in an alternative
front-end "harness" to the base Import UI/controller. This allows
employing the substantial work done to refine the Import system and
exploit its job-running capabilities while still building a more
purpose-built UI for the Frenkel lab.